+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+- Merge directives from Scalar/Enum/Type/Input/Interface extension node into target node.
+
 ## v6.33.4
 
 ### Fixed